tutorflow/tutorbackend/requirements.txt
asgiref==3.6.0
Django==4.1.7
django-cors-headers==3.14.0
django-filter==22.1
djangorestframework==3.14.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==5.2.2
gunicorn==20.1.0
numpy==1.24.3
pandas==2.0.1
pip-chill==1.0.1
psycopg2-binary==2.9.5
py-bcrypt==0.4
PyJWT==1.7.1
python-dateutil==2.8.2
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tzdata==2023.3
